// app/src/interfaces/form-fields-props.ts
export interface FormFieldsProps {
  firstName: string;
  lastName: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  email: string;
  phone: string;
  bedroomType: string;
  priceRange: string;
  howHear: string;
  isBroker: string;
  clientName: string;
  brokerageFirm: string;
  notes: string;
}